+++ Source/Objects/Hardware/GPIB/Pulser33220/ORPulser33220Model.h
@@ -74,8 +74,8 @@
 - (int) connectionProtocol;
 - (void) setConnectionProtocol:(int)aConnectionProtocol;
 
-- (unsigned long) vendorID;
-- (unsigned long) productID;
+- (NSUInteger) vendorID;
+- (NSUInteger) productID;
 - (NSString*) usbInterfaceDescription;
